MAJOR OVERHAUL
ditched the whole concept of static, CDEFO is now able to be instantiated. Otherwise, I wouldn't have been able to add concurrency.

I might have to revert all of this, but it seems to work alright enough. the only issue appears to be the fact that I'm doing so many String operations, so I have to work on transforming those into byte/char arrays so that I don't run out of RAM.