Another Major Overhaul
Strings to character arrays
Static is now in a fully functional state and object is broken since it needs to be converted to char arrays as well.
I made another light function, "Breathe," and I made chase more entertaining. Now all of the lighting starts in the center and moves outwards, which will look realy cool once I have a few meters of this stuff setup. I also made a "start_up" light function with a default mood-lighting. I'll probably also make an idle lighting function and perhaps include a little sound cue just to mkae it seem as interactive as possible.

The coding for the python scripts for video recording and audio recording are underway right now, as well as the continued developement of both current user experiences.